requirements.txt
streamlit
pyyaml
pydantic
pandas